+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-- Fixes issue where storage emulator throws an error due to non-standard whitespaces in filenames (#6834).
+- Data Connect emulator now reuses existing instances if they are emulating the same service.
+- Fixes issue where `deploy --only dataconnect` would error out with `cannot read property undefined`.
+- Fixes compatibility issue when uploading native symbols to Crashlytics on Intel-based Macs.